requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors
Flask-Bcrypt
pycryptodome
oauthlib
requests
supabase